JD: Changed error in Thread.Sleep which required a long as parameter

+++ ucsb-cs56-where-is-every1-Front-End/src/main/java/com/whereisevery1/database/model/Scraping.java
@@ -106,7 +106,7 @@ public class Scraping {
 			}
 
 			// random delay
-			Thread.sleep(r);
+			Thread.sleep((long) r);
 		}
 	}
 }
